test already showed a bug
